I've changed three files in PhetCommon to simplify piccolo integration.  The changes involve generalizing code so that I can successfully extend Module to PiccoloModule (which uses a PhetPCanvas instead of ApparatusPanel).  I have replaced the ContentPanel and TabbedApparatusPanelContainer code to use Module.getSimulationPanel(), so that in PiccoloModule I can return the PhetPCanvas.  In Module, the normal ApparatusPanel is returned (as per usual).  I also had to expose part of clockTicked() in Module, so that I could override it in PiccoloModule.
I'm happy to discuss these changes (or if there is a better way to do this), please let me know if you hit troubles.  Detailed changes follow:

in TabbedApparatusPanelContainer,
one call to getApparatusPanel() was replaced with getSimulationPanel()

in ContentPanel
two calls to getApparatusPanel() were replaced with getSimulationPanel()

in Module:
handleUserInput() was factored out of clockTicked(), and made protected (for overriding in PiccoloModule).
added getSimulationPanel(), which returns getApparatusPanel().