Minor changes to PhetApplicationConfig:
(1) If a flavor name is not specified, default the flavor name to the project name.

(2) Changed the syntax for standard properties to be consistent in both project properties
and localization properties files.  In both cases, the flavor name (project name, if no flavor
is specified) must be prefixed on the property name.  E.g. reactions-and-rates.version.major.