resolves #25
I have another layout comment but I was not there for the discussion of layout so am probably not aware of why things are as they are. I would rearrange the controls a little and try to get that area to look a little more like a control panel. I’d place the controls in the following order:

   1. I would have Force of Brake as is.
   2. I would have the vector items next. I’d remove the label Show Vectors and have the three check boxes in a row. “Show velocity vectors”, “Show acceleration Vectors”, and “Show Vector Components”
   3. “Ruler” checkbox remove the word Show.
   4. The last item I’d have is “allow non-tangential forces”. I would like to discuss rewording this to something simpler like “allow force in any direction” but that’s long. Unless the word tangential is part of the learning goals. Then we should keep it. Otherwise, I’d guess most will not play with it.
   5. Reset all and Help at the bottom.

This layout should have more symmetry and the fonts won’t vary and alignment should be the same. What do you think?