website: changed the standard header in site-utils.php to inject Javascript and to selectively check input elements
This is a quick fix which brings the Javascript initialization of the browse page from about 3 seconds to a few milliseconds (on my machine).

Currently there is Javascript code that checks every element and if it is an input element of an appropriate type, assigns a regex pattern for validation (and does a validation).  On the browse.php page, which has about 200+ simulations, this really adds up and only 10 of them actually need the check.

So now there is a bool that if true (the default), it will perform these checks.  If it is false, then it will skip the majority of them.  NOT a great solution, but a fast one until a better method can be implemented.

This has been developed specifically for browse.php and won't affect positively or negatively affect the performance on other pages.

TODO: Remove this Javascript initialization entirely and preprocess all the input elements in PHP.  This should be significantly faster, but will be more involved as it touches more code paths.