website: Speed improvements (many changes)
Server side:
The speed of many pages, mostly in the activities database but also some simulations, had started to suffer due to database getting big.  Most all of the data was formatted for html even though it was not being rendered.  Caching helps some but not entirely.  Now text is formatted only if it going to be rendered.  *Note: not 100% of all formatting has been removed.  Time is short and I'll have to ask for an extension.

There were many eval(dump_array_into_variables) which was irritating and slowed some pages down considerably.  Some of these as I came across them have been removed.  Would be nice to remove more.

Client side:
The loading speed of many pages, particularly those with lots of content, was slow.  There was much unnecessary JavaScript processing that looked at most every page element, when it was only looking for certain form input elements.  This problem was indepentant of the web page cache.