website: Removed need for "admin/uploads" directory, teacher's guides are now handled like contributions, removed URL to animated gif.
o  Teacher's guides are stored in the database and NOT in the website directory structure.  This was the original design intent and was not fully realized.  This will help with clutter in the web dirs, and make file handling more consistent.
o  Specifying an animated gif URL is old functionality and has been removed for some time.  Removed the reference from the edit sim page.