website: Fixed translation table script crashing when a language thumnail is not presnt
Solution:
We will generate one in plain text and save that image.  It will look different than the others, suggesting that someone needs to render a new version, both horizontal and vertical.
Currently the server does not support image rotate, and there is a check for this to stop the script from crashing.  If imagerotate is not available, it will just store the horizontal version.  This makes the table look bad, but again it is an indicator that something needs to be fixed.  At least the table renders.