Remove bogus background colors from flash-launcher html template.
There are 2 colors specified; the movie's background, and the browser background.
Ideally we'd like them to be the same, but we don't have access to the color, unless we either parse Mike's HTML files, or add the color to the sim's properties file.  

The current solution let's the movie set the background color, and leaves the browser background color alone -- which typically results in a thin white border around the movie.