First step in refactoring Sims to be compatible with post-IOM.  Major refactor involving changing sim inteface to be a class rather than just an associative array and a supporting set of functions in sim-utils.php.
simulations/sims.php has been refactored to use this new class
exclusively.  Everything else still relies on sim-utils.php.

This is to save my place amist this complicated change.  All the code
works as one would expect but interanally it is straddling 2 different
methods.

Next step will be do add backwards (seamless?) compatibility with
pre-IOM as a subclass that can be dropped when it is no longer
needed.  Once that is complete, the rest of the scripts will be
refactored to use the newer, simpler, (faster?) method.