Graphics created by Noah P for this sim.  Noah said:
Here they are...but they are Illustrator CS5, so not sure about compatibility. I also saved as SVG, but not sure that will keep the 3D effects that I used.

Please note that the transparent area in the charge meter was done in Photoshop. That area is not transparent in the Illustrator file, only the png that I produced.